Seems final code for testing is added (miner chnages for time in serial.cpp for all contracts is pending

+++ 2.Ballot/Serial/Util/FILEOPR.h
@@ -21,17 +21,29 @@ class FILEOPR
 	int getRFunC( int nCFun );
 		
 	
-	void getInp(int* nProposal, int* nVoter,
-				int *nThreads, int* nAUs, double* lemda);
+	void getInp(int* nProposal,
+	            int* nVoter,
+	            int *nThreads,
+	            int* nAUs,
+	            double* lemda);
 	
-	void writeOpt( int nProposal, int nVoter, int nThreads, 
-					int nAUs, double TTime[], float_t mTTime[], 
-					float_t vTTime[], int aCount[], int vAUs, 
-					list<double>&mIT, list<double>&vIT );
+	void writeOpt(int nProposal,
+	              int nVoter,
+	              int nThreads, 
+	              int nAUs,
+	              double TTime[],
+	              float_t mTTime[], 
+	              float_t vTTime[],
+	              int aCount[],
+	              int vAUs,
+	              list<float>&mIT,
+	              list<float>&vIT);
 	
-	void genAUs(int numAUs, int numVoter, int numProposal,
-				int nFunC, vector<string>& ListAUs
-				);
+	void genAUs(int numAUs,
+	            int numVoter,
+	            int numProposal,
+	            int nFunC,
+	            vector<string>& ListAUs);
 	
 	void printCList(int AU_ID, list<int>&CList);
